+++ src/custom-nodes/string-array-nodes/StringArrayCallNode.ts
@@ -7,9 +7,12 @@ import { TIdentifierNamesGeneratorFactory } from '../../types/container/generato
 import { TStatement } from '../../types/node/TStatement';
 import { TStringArrayIndexNodeFactory } from '../../types/container/custom-nodes/string-array-index-nodes/TStringArrayIndexNodeFactory';
 
+import { IArrayUtils } from '../../interfaces/utils/IArrayUtils';
 import { ICustomCodeHelperFormatter } from '../../interfaces/custom-code-helpers/ICustomCodeHelperFormatter';
 import { IOptions } from '../../interfaces/options/IOptions';
 import { IRandomGenerator } from '../../interfaces/utils/IRandomGenerator';
+import { IStringArrayScopeCallsWrapperParameterIndexesData } from '../../interfaces/node-transformers/string-array-transformers/IStringArrayScopeCallsWrapperParameterIndexesData';
+import { IStringArrayStorage } from '../../interfaces/storages/string-array-transformers/IStringArrayStorage';
 
 import { initializable } from '../../decorators/Initializable';
 
@@ -44,9 +47,17 @@ export class StringArrayCallNode extends AbstractStringArrayCallNode {
     private stringArrayCallsWrapperName!: string;
 
     /**
+     * @type {IStringArrayScopeCallsWrapperParameterIndexesData | null}
+     */
+    @initializable()
+    private stringArrayCallsWrapperParameterIndexesData!: IStringArrayScopeCallsWrapperParameterIndexesData | null;
+
+    /**
      * @param {TIdentifierNamesGeneratorFactory} identifierNamesGeneratorFactory
      * @param {TStringArrayIndexNodeFactory} stringArrayIndexNodeFactory
      * @param {ICustomCodeHelperFormatter} customCodeHelperFormatter
+     * @param {IStringArrayStorage} stringArrayStorage
+     * @param {IArrayUtils} arrayUtils
      * @param {IRandomGenerator} randomGenerator
      * @param {IOptions} options
      */
@@ -56,6 +67,8 @@ export class StringArrayCallNode extends AbstractStringArrayCallNode {
         @inject(ServiceIdentifiers.Factory__IStringArrayIndexNode)
             stringArrayIndexNodeFactory: TStringArrayIndexNodeFactory,
         @inject(ServiceIdentifiers.ICustomCodeHelperFormatter) customCodeHelperFormatter: ICustomCodeHelperFormatter,
+        @inject(ServiceIdentifiers.IStringArrayStorage) stringArrayStorage: IStringArrayStorage,
+        @inject(ServiceIdentifiers.IArrayUtils) arrayUtils: IArrayUtils,
         @inject(ServiceIdentifiers.IRandomGenerator) randomGenerator: IRandomGenerator,
         @inject(ServiceIdentifiers.IOptions) options: IOptions
     ) {
@@ -63,6 +76,8 @@ export class StringArrayCallNode extends AbstractStringArrayCallNode {
             identifierNamesGeneratorFactory,
             stringArrayIndexNodeFactory,
             customCodeHelperFormatter,
+            stringArrayStorage,
+            arrayUtils,
             randomGenerator,
             options
         );
@@ -70,17 +85,20 @@ export class StringArrayCallNode extends AbstractStringArrayCallNode {
 
     /**
      * @param {string} stringArrayCallsWrapperName
+     * @param {IStringArrayScopeCallsWrapperParameterIndexesData | null} stringArrayCallsWrapperParameterIndexesData
      * @param {number} index
      * @param {number} indexShiftAmount
      * @param {string | null} decodeKey
      */
     public initialize (
         stringArrayCallsWrapperName: string,
+        stringArrayCallsWrapperParameterIndexesData: IStringArrayScopeCallsWrapperParameterIndexesData | null,
         index: number,
         indexShiftAmount: number,
         decodeKey: string | null
     ): void {
         this.stringArrayCallsWrapperName = stringArrayCallsWrapperName;
+        this.stringArrayCallsWrapperParameterIndexesData = stringArrayCallsWrapperParameterIndexesData;
         this.index = index;
         this.indexShiftAmount = indexShiftAmount;
         this.decodeKey = decodeKey;
@@ -90,21 +108,46 @@ export class StringArrayCallNode extends AbstractStringArrayCallNode {
      * @returns {TStatement[]}
      */
     protected getNodeStructure (): TStatement[] {
-        const callExpressionArgs: ESTree.Expression[] = [
-            this.getStringArrayIndexNode(this.indexShiftAmount + this.index)
-        ];
+        const resultIndex: number = this.indexShiftAmount + this.index;
 
-        if (this.decodeKey) {
-            callExpressionArgs.push(this.getRc4KeyLiteralNode(this.decodeKey));
+        const indexNode: ESTree.Expression = this.getStringArrayIndexNode(resultIndex);
+        const rc4KeyLiteralNode: ESTree.Literal | null = this.decodeKey
+            ? this.getRc4KeyLiteralNode(this.decodeKey)
+            : null;
+
+        // filling call expression arguments with a fake arguments first
+        const callExpressionArgs: ESTree.Expression[] = this.arrayUtils.fillWithRange(
+            !this.stringArrayCallsWrapperParameterIndexesData
+                // root string array calls wrapper
+                ? AbstractStringArrayCallNode.stringArrayRootCallsWrapperParametersCount
+                // scope string array calls wrapper
+                : this.options.stringArrayWrappersParametersMaxCount,
+            () => this.getFakeStringArrayIndexNode(resultIndex)
+        );
+
+        callExpressionArgs.splice(
+            this.stringArrayCallsWrapperParameterIndexesData?.valueIndexParameterIndex ?? 0,
+            1,
+            indexNode
+        );
+
+        if (this.stringArrayCallsWrapperParameterIndexesData) {
+            callExpressionArgs.splice(
+                this.stringArrayCallsWrapperParameterIndexesData.decodeKeyParameterIndex,
+                1,
+                // use rc4 key literal node if exists or a node with fake string array wrapper index
+                rc4KeyLiteralNode ?? this.getFakeStringArrayIndexNode(resultIndex)
+            );
+        } else if (rc4KeyLiteralNode) {
+            callExpressionArgs.splice(1, 1, rc4KeyLiteralNode);
+        } else {
+            // have to delete element
+            callExpressionArgs.splice(1, 1);
         }
-
-        const stringArrayIdentifierNode: ESTree.Identifier =
-            NodeFactory.identifierNode(this.stringArrayCallsWrapperName);
-
 
         const structure: TStatement = NodeFactory.expressionStatementNode(
             NodeFactory.callExpressionNode(
-                stringArrayIdentifierNode,
+                NodeFactory.identifierNode(this.stringArrayCallsWrapperName),
                 callExpressionArgs
             )
         );
@@ -112,5 +155,28 @@ export class StringArrayCallNode extends AbstractStringArrayCallNode {
         NodeUtils.parentizeAst(structure);
 
         return [structure];
+    }
+
+    /**
+     * @param {number} actualIndex
+     * @returns {Expression}
+     */
+    private getFakeStringArrayIndexNode (actualIndex: number): ESTree.Expression {
+        return this.getStringArrayIndexNode(this.getFakeStringArrayIndex(actualIndex));
+    }
+
+    /**
+     * @param {number} actualIndex
+     * @returns {number}
+     */
+    private getFakeStringArrayIndex (actualIndex: number): number {
+        const stringArrayStorageLength: number = this.stringArrayStorage.getLength();
+
+        const fakeIndexOffset: number = stringArrayStorageLength / 2;
+
+        const minimumIndex: number = actualIndex - fakeIndexOffset;
+        const maximumIndex: number = actualIndex + fakeIndexOffset;
+
+        return this.randomGenerator.getRandomInteger(minimumIndex, maximumIndex);
     }
 }