+++ src/interfaces/node-transformers/string-array-transformers/IStringArrayScopeCallsWrapperParameterIndexesData.ts
@@ -1,0 +1,11 @@
+export interface IStringArrayScopeCallsWrapperParameterIndexesData {
+    /**
+     * @type {number}
+     */
+    valueIndexParameterIndex: number;
+
+    /**
+     * @type {number}
+     */
+    decodeKeyParameterIndex: number;
+}